+++ mainproject/src/main/scala/com/raphtory/testCases/wordSemantic/graphbuilders/CoMatGB.scala
@@ -26,7 +26,7 @@ class CoMatGB extends GraphBuilder[String] {
           occurenceTime,
           srcClusterId,
           dstClusterId,
-          Properties(LongProperty("Frequency", coocWeight))
+          Properties(DoubleProperty("Frequency", coocWeight))
         )
       }
